% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spiges helpers.R
\name{check_spiges_tables}
\alias{check_spiges_tables}
\title{spiges helpers
Check presence of required SpiGes tables}
\usage{
check_spiges_tables(spiges_data, tablenames)
}
\arguments{
\item{spiges_data}{list or object of class 'spiges_data'}

\item{tablenames}{character vector of table names to check}
}
\value{
invisible(NULL) if all tables present, otherwise throws an error
}
\description{
spiges helpers
Check presence of required SpiGes tables
}
