% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spiges helpers.R
\name{check_spiges_var}
\alias{check_spiges_var}
\title{Check presence of required SpiGes variables}
\usage{
check_spiges_var(spiges_table, var_names)
}
\arguments{
\item{spiges_table}{dataframe of a SpiGes-data Object, eg. `admin`}

\item{var_names}{character vector of variable names to check}
}
\value{
invisible(NULL) if all tables present, otherwise throws an error
}
\description{
Check presence of required SpiGes variables
}
