% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spiges_data-filter.R
\name{filter.spiges_data}
\alias{filter.spiges_data}
\title{Filter spiges_data by admin table}
\usage{
filter.spiges_data(.data, ...)
}
\arguments{
\item{.data}{A data frame, data frame extension (e.g. a tibble), or a
lazy data frame (e.g. from dbplyr or dtplyr). See \emph{Methods}, below, for
more details.}

\item{...}{<\code{\link[rlang:args_data_masking]{data-masking}}> Expressions that
return a logical value, and are defined in terms of the variables in
\code{.data}. If multiple expressions are included, they are combined with the
\code{&} operator. Only rows for which all conditions evaluate to \code{TRUE} are
kept.}
}
\description{
Filter spiges_data by admin table
}
