% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmt_splg.R
\name{fmt_splg}
\alias{fmt_splg}
\title{Format SpiGes data to a format for SPLG-Grouper Input}
\usage{
fmt_splg(
  spiges_data,
  type = c("group", "control", "output"),
  format = c("TEXT", "XML", "JSON"),
  version = "1.4"
)
}
\arguments{
\item{spiges_data}{A list of data frames representing the SpiGes data structure}

\item{type}{One of 'TEXT', 'XML', 'JSON', for defining a input format for the SPLG-Grouper}

\item{version}{SpiGes version (default: "1.4")}
}
\value{
character vector, prepared for writing to directory `1-input` of SPLG-Grouper
}
\description{
This function prepares SpiGes data so they can be used as input for the SPLG-Grouper.
The data are transformed according to the SPLG-Grouper specification.
}
\examples{
\dontrun{
fmt_splg(spiges_data)
}
}
