% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmt_splg.R
\name{fmt_splg_proc}
\alias{fmt_splg_proc}
\title{Internal: format procedure (CHOP) records for SPLG-Grouper input (supports TEXT and JSON outputs)}
\usage{
fmt_splg_proc(proc, format = c("TEXT", "XML", "JSON"), version = "1.4")
}
\description{
Internal: format procedure (CHOP) records for SPLG-Grouper input (supports TEXT and JSON outputs)
}
