% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmt_swissdrg.R
\name{fmt_swissdrg}
\alias{fmt_swissdrg}
\title{Export SpiGes data to SwissDRG BATCH_2017 format}
\usage{
fmt_swissdrg(
  spiges_data,
  format = "BATCH_2017",
  tariff = c("SwissDRG", "ST Reha", "TARPSY")
)
}
\arguments{
\item{spiges_data}{A list of data frames representing the SpiGes data structure}

\item{version}{SpiGes version (default: "1.4")}
}
\value{
character vector, prepared for writing to directory of SwissDRG-Grouper
}
\description{
This function exports SpiGes data to the SwissDRG BATCH_2017 format.
}
\examples{
\dontrun{
fmt_swissdrg(spiges_data)
}
}
