% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmt_swissdrg.R
\name{fmt_swissdrg_babydata}
\alias{fmt_swissdrg_babydata}
\title{Internal: format babydata records for SwissDRG-Grouper input}
\usage{
fmt_swissdrg_babydata(admin, neugeb)
}
\description{
Internal: format babydata records for SwissDRG-Grouper input
}
