% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmt_swissdrg.R
\name{fmt_swissdrg_diag}
\alias{fmt_swissdrg_diag}
\title{Internal: format diagnoses (ICD) records for SwissDRG-Grouper input}
\usage{
fmt_swissdrg_diag(diag)
}
\description{
Internal: format diagnoses (ICD) records for SwissDRG-Grouper input
}
