% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmt_swissdrg.R
\name{fmt_swissdrg_los}
\alias{fmt_swissdrg_los}
\title{Internal: format length-of-stay records for SwissDRG-Grouper input}
\usage{
fmt_swissdrg_los(los, tariff)
}
\description{
Internal: format length-of-stay records for SwissDRG-Grouper input
}
