% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmt_swissdrg.R
\name{fmt_swissdrg_medi}
\alias{fmt_swissdrg_medi}
\title{Internal: format medication records for SwissDRG-Grouper input}
\usage{
fmt_swissdrg_medi(medi)
}
\description{
Internal: format medication records for SwissDRG-Grouper input
}
