% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmt_swissdrg.R
\name{fmt_swissdrg_proc}
\alias{fmt_swissdrg_proc}
\title{Internal: format procedure (CHOP) records for SwissDRG-Grouper input}
\usage{
fmt_swissdrg_proc(proc)
}
\description{
Internal: format procedure (CHOP) records for SwissDRG-Grouper input
}
