% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_xml.R
\name{import_spiges_csv}
\alias{import_spiges_csv}
\title{Import SpiGes data from distributed CSV files}
\usage{
import_spiges_csv(directory, version = "1.4")
}
\arguments{
\item{directory}{Path to the directory containing the CSV files}

\item{version}{SpiGes version (default: "1.4")}
}
\value{
A list of data frames representing the SpiGes data structure
}
\description{
This function imports SpiGes data from distributed CSV files (12 files)
and converts it to a structured R object. All variables are converted to their
appropriate formats as defined in the SpiGes documentation.
}
\examples{
\dontrun{
data <- import_spiges_csv("path/to/csv/directory")
}
}
