% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spiges_data-class.R
\name{problems.spiges_data}
\alias{problems.spiges_data}
\title{Extract parsing and key problems from spiges_data}
\usage{
problems.spiges_data(x, ...)
}
\arguments{
\item{x}{A spiges_data object}

\item{...}{unused}
}
\description{
Extract parsing and key problems from spiges_data
}
