% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_csv.R
\name{read_bfs_spiges}
\alias{read_bfs_spiges}
\title{Read BFS hospital statistics (SpiGes) from a directory}
\usage{
read_bfs_spiges(dirname, tables = NULL, col_types = NULL, version = "1.4")
}
\arguments{
\item{dirname}{Directory containing the CSV files.}

\item{tables}{Character vector of table names to read (fuzzy). If \code{NULL} (default), all available tables in the directory
are read.}

\item{col_types}{Optional \code{readr::cols_only()} object that is used as
\code{col_types} for all tables (overrides table-specific col_types).}

\item{version}{= '1.4' SpiGes Version.}
}
\value{
A list with the requested tables (always containing \code{admin}),
  with attributes \code{Datenjahr}, \code{Datenversion}, \code{Source}.
}
\description{
Read BFS hospital statistics (SpiGes) from a directory
}
