% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_Spiges_xml.R
\name{read_spiges_xml}
\alias{read_spiges_xml}
\title{Import SpiGes data from XML files}
\usage{
read_spiges_xml(data_file, id_file, version = "1.4")
}
\arguments{
\item{data_file}{Path to the SpiGes data XML file}

\item{id_file}{Path to the SpiGes identifiers XML file}

\item{version}{SpiGes version (default: "1.4")}
}
\value{
A list of data frames representing the SpiGes data structure
}
\description{
This function imports SpiGes data from XML files (data file and identifiers file)
and converts it to a structured R object. All variables are converted to their
appropriate formats as defined in the SpiGes documentation.
}
\examples{
\dontrun{
data_file <- system.file("extdata", "do-d-14.04-SPIGES-2024-01.xml", package = "SpiGesR")
id_file <- system.file("extdata", "do-d-14.04-SPIGES-2024-03.xml", package = "SpiGesR")
data <- import_spiges_xml(data_file, id_file)
}
}
