% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spiges helpers.R
\name{spiges2date}
\alias{spiges2date}
\title{coerce variable of different classes to Date (withour Hour)}
\usage{
spiges2date(x)
}
\arguments{
\item{x}{vector to be coerced to SpiGes-Date}
}
\value{
vector fo class Date
}
\description{
coerce variable of different classes to Date (withour Hour)
}
