% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spiges helpers.R
\name{spiges2datestr}
\alias{spiges2datestr}
\title{coerce variable of different classes to character vector date string (withour Hour)}
\usage{
spiges2datestr(x)
}
\arguments{
\item{x}{vector to be coerced to character (format SpiGes-Date)}
}
\value{
vector fo class character
}
\description{
coerce variable of different classes to character vector date string (withour Hour)
}
